Modificando o tipo do atributo preço

+++ GestaoDeEstoque/Models/Produto.cs
@@ -9,7 +9,7 @@ namespace GestaoDeEstoque.Models
     {
         public int Id { get; set; }
         public string Nome { get; set; }
-        public decimal Preco { get; set; }
+        public float Preco { get; set; }
         public int Quantidade { get; set; }
     }
 }
